# Fixes [SUREFIRE-1516]: Poor performance in reuseForks=false