--- a/surefire-api/src/main/java/org/apache/maven/surefire/booter/CommandReader.java
+++ b/surefire-api/src/main/java/org/apache/maven/surefire/booter/CommandReader.java
@@ -412,6 +412,10 @@ public final class CommandReader
                                 CommandReader.this.wakeupIterator();
                                 callListeners( command );
                                 break;
+                            case BYE_ACK:
+                                //After SHUTDOWN no more commands can come. Hence, do NOT go back to blocking in IO
+                                callListeners( command );
+                                return;
                             default:
                                 callListeners( command );
                                 break;