--- a/surefire-api/src/main/java/org/apache/maven/surefire/booter/CommandReader.java
+++ b/surefire-api/src/main/java/org/apache/maven/surefire/booter/CommandReader.java
@@ -413,7 +413,7 @@ public final class CommandReader
                                 callListeners( command );
                                 break;
                             case BYE_ACK:
-                                //After SHUTDOWN no more commands can come. Hence, do NOT go back to blocking in IO
+                                //After SHUTDOWN no more commands can come. Hence, do NOT go back to blocking in I/O
                                 callListeners( command );
                                 return;
                             default: