@Tibor17 2019-11-12
The build fails on macOS.
`ForkModeMultiModuleIT.testForkCountTwoNoReuse:70->doTest:137 » SurefireVerifier`
Can you pls have a look?
https://github.com/apache/maven-surefire/pull/253/checks?check_run_id=298330219