@Tibor17 2019-12-17
The build failed again https://github.com/apache/maven-surefire/runs/352389395
@jon-bell What happens if you push a new dummy commit several time an hour?