@jon-bell 2019-12-17
@tibor17 I made a dummy commit and it looked like it passed - https://github.com/apache/maven-surefire/commit/06c74da235fe9511b1452df54a74449e92e4c921