@Tibor17 2019-12-18
@jon-bell 
ok, let's continue with a new commit for https://github.com/apache/maven-surefire/pull/253#discussion_r358743737
and then i will create a branch to run the build on our Jenkins.