@Tibor17 2019-12-29
@jon-bell Thx for contributing. I used your changes, updated the IT 705 and closed the Jira as fixed.
Can you pls perform new measurements with master, how far we are from the last measurement?