@jon-bell 2019-12-30
@Tibor17 - Thanks! The numbers from current master (5148b02ba552cd79ac212b869dec10d01ba4d2e6) look pretty much the same on my side (down to 16.9msec to tear down the process).